prefer background-color to background-image
and refactor by extracting code inside a ShelfDot component
